split test fixtures from setup.tsx
