+++ enterprise/frontend/src/embedding-sdk-package/hooks/public/use-metabase-query/use-metabase-query-hooks.unit.spec.tsx
@@ -2,13 +2,13 @@
 
 import {
   TEST_DATASET_QUERY,
-  TEST_SCHEMA,
   createDeferred,
   mbqlQuery,
   mockUseLazySelector,
   resetTestState,
   stubSdkBundle,
-} from "./setup";
+} from "./tests/setup";
+import { TEST_SCHEMA } from "./tests/fixtures";
 
 import { act, renderHook, waitFor } from "@testing-library/react";
